create JobListingsModal
